ftr:Agrendo ajustes product controller

+++ src/controllers/products.controller.js
@@ -4,7 +4,7 @@ const createProduct = async (req, res, next) => {
   try {
     const newProduct = new Product(req.body);
     await newProduct.save();
-    res.status(200).json({ msg: "Product Created" });
+    res.status(201).json({ message: "Product Created" });
   } catch (error) {
     next(error);
   }
@@ -13,6 +13,8 @@ const createProduct = async (req, res, next) => {
 const getProducts = async (req, res, next) => {
   try {
     const products = await Product.find();
+    if (!products)
+      return res.status(204).json({ message: "No products found" });
     res.status(200).json(products);
   } catch (error) {
     next(error);
@@ -21,8 +23,15 @@ const getProducts = async (req, res, next) => {
 
 const getProduct = async (req, res, next) => {
   const { id } = req.params;
+
+  if (!id) return res.status(400).json({ message: "Product Id required" });
+
   try {
     const productFind = await Product.findById(id);
+
+    if (!productFind)
+      return res.status(404).json({ message: "No product matches Id" });
+
     res.status(200).json(productFind);
   } catch (error) {
     next(error);
@@ -31,6 +40,9 @@ const getProduct = async (req, res, next) => {
 
 const updateProduct = async (req, res, next) => {
   const { id } = req.params;
+
+  if (!id) return res.status(400).json({ message: "Id parameter required" });
+
   try {
     await Product.findByIdAndUpdate(id, req.body);
     res.status(200).json({ msg: "Product Update" });
@@ -41,12 +53,21 @@ const updateProduct = async (req, res, next) => {
 
 const deleteProduct = async (req, res, next) => {
   const { id } = req.params;
+
+  if (!id) return res.status(400).json({ message: "Id parameter required" });
+
   try {
     await Product.findByIdAndRemove(id);
-    res.status(200).json({ msg: "Product Deleted" });
+    res.status(200).json({ message: "Product Deleted" });
   } catch (error) {
     next(error);
   }
 };
 
-module.exports = { createProduct, getProducts, getProduct, updateProduct, deleteProduct };
+module.exports = {
+  createProduct,
+  getProducts,
+  getProduct,
+  updateProduct,
+  deleteProduct,
+};
